requirements.txt
flask
torch
transformers
datasets
openai-whisper
ffmpeg-python
pydub
soundfile
librosa
pyttsx3
python-multipart
